Move styling injection into its own file
