+++ src/ui/index.js
@@ -1,0 +1,1 @@
+export {default as injectStyling} from "./styling";
